+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Migrate proxy to Diamond pattern (ERC-2535):
   - Remove ENS module (#225)
+  - Add Diamond contract unit tests (#224)
   - Fix `fallback` and `receive` (#223)
   - Migrate contracts (#222)